requirements.txt
alembic==1.10.3
annotated-types==0.5.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
oauthlib==3.2.2
plyer==2.1.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.10
pydantic_core==2.1.2
PyMySQL==1.1.0
pyparsing==3.0.9
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
redbird==0.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rocketry==2.5.1
rsa==4.9
shiboken6==6.5.2
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.15
